CHORE : 기존 PlaceReviewModule을 MainModule로 옮기기
